Bugfix: fix lsp#utils#path_to_uri for windows
On windows, the path for the current buffer is created incorrectly.
When we path the a:path through s:encode_uri, it garbles the c:/
part of the path and uri encodes it.

In order to fix this bug, I changed the function to prepend 'file:///'
to the front of the path that has slashes switched to forward slashes.