Fix lsp#utils#path_to_uri for windows
When encoding the buffer path on Windows the path_to_uri transformed the
C: to C%3A causing problems to LSP functions.

To solve this it as necessary to change the function encode_uri to
receive the start position to encode, so the functions that call it can
specify were to start the encoding.