Fix bug in apply_text_edits method.
Text edits have to be applied from the end of the document, otherwise
the earlier edit would invalidate position of all other edits.

Example input: "for ( )"

Text Edits: Remove spaces - 3rd and 5th character

If we just apply the edits in order, we would first remove the space,
but then if we go to 5th character in this line, we would remove ')' and
we would end up with "for( " instead of "for()".

I found this bug when playing with typescript LSP and trying to format a
document - I saw strange behavior of some characters disappearing
exactly as in the example above.