Reapply fixes for EOL injection.
PRs #160 (merged in 3cee859) and #158 (merged in 9806023) introduce the
use of a platform-specific EOL marker and logic to conditionally inject
an extra EOL at the end of the last line when sending file contents to
the language server.

These fixes resolved an issue where some backends (pyls with flake8, for
example) would raise a linter error due to a missing EOL at the end of
the file. They were reverted in 34c2a1e and 1e0c19d, presumably by
accident.