Adds `g:lsp_omni_completion_ignorecase` option to indicates whether or not ignorecase to matchs completion word
In most cases, ignorecase is not necessary, so we may add an option to control whether to turn it on.
Defaults, `g:lsp_omni_completion_ignorecase` is `0`, meaning case sensitive, if `g:lsp_omni_completion_ignorecase`
is set to `1`, or set `ignorecase` options(eg. `:set ignorecase`), it will turn case insensitive on.