Check for codeActionKinds in codeActionProvider (#382)
This was tested with ccls, which will return CodeActionOptions instead
of a boolean in codeActionProvider. According to spec this is an invalid
response by ccls, because as a client we didn't signal
codeActionLiteralSupport. But I think it can be safely assumed that the
server provides code actions when he returns a non empty list of
codeActionKinds.