Use "P" instead of "p" to insert text
When running `:LspDocumentFormat` on a Java file, it rewrote the imports
of this source:

    package my_package

    import java.util.ArrayList;
    import java.util.List;

    public class MyClass {
        // ...
    }

to this:

    package my_package

    import java.util.ArrayList;
    import java.util.List;
    public class MyClass {
        // ...
    }

I first dug in to the implementation of `textDocument/formatting` on the
Java server side, and it appeared to be calculating the location to
insert the text correctly, yet it was being pushed down an extra line.

With this change, the new set of imports gets inserted at the correct
location.