Add support for client-side commands
The eclipse.jdt.ls project implemented workspace edits using custom
client-side commands, apparently due to an ambiguity in the spec. As a
result, other clients have had to implement these custom commands.

The only one implemented in this commit is java.apply.workspaceEdit,
though there are other ones that can be added on an as-needed basis.

See
https://github.com/eclipse/eclipse.jdt.ls/issues/376#issuecomment-333923685