Support opening locations in new window
This patch adds support for opening locations (go to definition,
declaration, etc.) in a new window (split) vs the current window or a
preview window. I couldn't find a way to ask for this explicitly. This
patch extends all 4 goto commands to accept a `where` parameter which
which enables the user to select where to open the result. Options are
'split', 'vsplit', and 'preview'.