Load `retriggerCharacters` for signature help.
The LSP (v3.15) specification dictates both `triggerCharacters` and
`retriggerCharacters` in SignatureHelpOptions to trigger Signature
Help.  Currently, vim-lsp only loads `triggerCharacters` and not
`retriggerCharacters`.

This PR detects if the LSP server also sent `retriggerCharacters` along
`triggerCharacters`.  If so, we append them to the list of
`triggerChars` and treat them as if they were regular `triggerChars`.