Improved neovim floating windows a LOT
Now, all windows are positioned relative to the global grid. This fixes
many problems that I discovered while using this plugin. Also, I made
the float-positioning logic a lot simpler, and added options to 'dock'
the preview window (for completion). Also, now the floating windows have a
single-width padding to help differentiate between the floating text and
the main window text better.

Configuration options for specifing whether docking is to be enabled or
not has been added, and also the maximum height of the docked window.