Fix frozen completion menu in neovim
The completion menu was getting frozen due to the floating
documentation popup. Previously, it was implemented by first creating an
empty window, then entering it, filling the text in it, and then
resizing it accordingly before returning to the original window. This
messed with the completion popup when it was shown above the cursor
(due to less space below it).

This commit fixes that issue by first creating and filling a buffer,
then computing the required size and position, and then finally creating
the popup window (without entering it). Due to this major change in the
workflow, the existing window functions could not be used, and all the code
required to create the popup had to be accumulated into a single place
(documentation.vim)

Due to the fact that we never enter the floating window, all the changes
that need to be made to the window had to be implemented through the
nvim api funtions.

This commit also inclues a few miscellaneous changes that I had
forgotten to include in my previous two commits.

Also see https://github.com/prabirshrestha/vim-lsp/pull/921#issuecomment-714914185