Miscellaneous changes and fixes
'signcolumn' is property of window, not buffer. So, set it accordingly
using the window API function rather than the buffer API function.

Do not add additional spaces on both sides of buffer text, as it can
result in python auto-linters go mad over the unexpected indentation.
This may also apply for other languages that I an unaware of.
(Originally meant for effect of padding)