Improve placement of documentation popup in neovim
If the available space in either side of the completion popup is less
than 3/4 of the required width, then we shall place the documentation
popup below (or above, if necessary) of the completion popup.

Previously, the documentation was just squashed into whatever space was
available. It made documentation undreadable in some cases.