Add prefix to completions if start col changes.
`textEdit` completions can cause the column where the completion begins
to change. Current behaviour is to just take the minimum of such start
characters. This breaks autocompletion of edits that don't start at this
earliest position as:

(a) They are filtered out by `prefix` or `contains` autocomplete filters
as they no longer match the `last_typed_word`
(b) They insert at the wrong location even if the above is fixed.

The solution included here is to add any extra characters included by
the moving of the start column as a prefix to any insertion completions.