Send didClose on BufUnload
The didClose notification is a mandatory part of the LSP protocol,
without it, the language server might have a wrong view of which files
the client has opened. For example, rust-analyzer will provide a code
action to create a missing module, but will stop providing this action
once we have sent a didOpen notification for the corresponding file.
Without the didClose notification that code action will still be
missing even if we unload the buffer without ever writing it to disk.
But with this notification, the state is properly synced and the code
action reappears once we unload the buffer.