Add per-file LSP activation rules allowpaths, blockpaths
`allowpaths` and `blockpaths` are lists of regular expressions matched
with the full path of the active file.  LSP is enabled if it matches
any pattern in the `allowpaths` list or when `allowpaths` is not present
in the server_info dictionary.  LSP is disabled if any of the `blockpaths`
patterns matches the current buffer.  The `blockpaths` rule has priority
when both `allowpaths` and `blockpaths` match.

The `allowpaths` and `blockpaths` are evaluated only if the LSP is
enabled by the `allowlist`, `blocklist` rules for buffer filetypes.