Fix completion item text selection priority (#1614)
* Fix completion item text selection priority

According to the Language Server Protocol 3.17 specification:

- filterText is only used for filtering and sorting completion items
- When textEdit is provided, insertText must be ignored

* Fix filtering should be performed on filterText

If `filterText` exists, the filtering target has been changed from `word`.

The filter is defined as follows in the LSP specification:

```
A string that should be used when filtering a set of completion items.
When omitted the label is used as the filter text for this item.
```

* Fix remove unnecessary empty check condition

---------

Co-authored-by: mattn <mattn.jp@gmail.com>